+++ src/mtimer.c
@@ -1,6 +1,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <assert.h>
 
@@ -43,7 +44,7 @@ static void mTimer_destroy(mTimer *self)
 
 static BOOL mTimer_timerProc(HWND hwnd, int id, DWORD total_count)
 {
-	mTimer * timer = (mTimer*)id;
+	mTimer * timer = (mTimer*)(intptr_t)id;
 
 	if(timer)
 	{
@@ -69,7 +70,7 @@ static BOOL mTimer_start(mTimer *self)
 
 	if(IsWindow(hwnd))
 	{
-		return SetTimerEx(hwnd, (int)self, self->interval, (TIMERPROC)mTimer_timerProc);
+		return SetTimerEx(hwnd, (int)(intptr_t)self, self->interval, (TIMERPROC)mTimer_timerProc);
 	}
 
 	return FALSE;
@@ -79,7 +80,7 @@ static void mTimer_stop(mTimer *self)
 {
 	HWND hwnd = _c(self)->getParent(self);
 
-	KillTimer(hwnd, (int)self);
+	KillTimer(hwnd, (int)(intptr_t)self);
 }
 
 static BOOL mTimer_setProperty(mTimer* self, int id, DWORD value)
@@ -95,7 +96,7 @@ static BOOL mTimer_setProperty(mTimer* self, int id, DWORD value)
 		hwnd = _c(self)->getParent(self);
 		if(HWND_NULL == hwnd)
 			return FALSE;
-		isStart = IsTimerInstalled(hwnd, (int)self);
+		isStart = IsTimerInstalled(hwnd, (int)(intptr_t)self);
 		if(isStart){
 			_M(self, stop);
 			self->interval = value;
